# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mithra AI Family Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/me/settings:
    get:
      operationId: getUserSettings
      tags: [users]
      summary: Get user settings
      responses:
        "200":
          description: User settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
    patch:
      operationId: updateUserSettings
      tags: [users]
      summary: Update user settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"

  # ─── Chats ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, clerkId, email, name, role, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, member]
        isActive:
          type: boolean
        createdAt:
          type: string
    UserProfileUpdate:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
    UserSettings:
      type: object
      required: [id, userId]
      properties:
        id:
          type: integer
        userId:
          type: integer
        theme:
          type: string
          enum: [dark, light, system]
        accentColor:
          type: string
        language:
          type: string
        animationsEnabled:
          type: boolean
        voiceEnabled:
          type: boolean
        voiceId:
          type: ["string", "null"]
        voiceSpeed:
          type: number
        voicePitch:
          type: number
        customInstructions:
          type: ["string", "null"]
        notificationsEnabled:
          type: boolean
    UserSettingsUpdate:
      type: object
      properties:
        theme:
          type: string
        accentColor:
          type: string
        language:
          type: string
        animationsEnabled:
          type: boolean
        voiceEnabled:
          type: boolean
        voiceId:
          type: ["string", "null"]
        voiceSpeed:
          type: number
        voicePitch:
          type: number
        customInstructions:
          type: ["string", "null"]
        notificationsEnabled:
          type: boolean

    # ─── Chat ────────────────────────────────────────────────────────────────
...